Fix up samples using new approach.
Narrative is now in the sample file itself, using triple-slashes.
