+++ .watchmanconfig
@@ -1,3 +1,3 @@
 {
-  "ignore_dirs": ["tst/dist"]
+  "ignore_dirs": ["tst/dist", "sam/dist"]
 }
