+++ sam/But.js
@@ -1,6 +1,20 @@
 import {U, E, T} from '../src/Up';
 var dom, config;
 
+/// ### Buttons and labels
+///
+/// Let's move on from timers and work with real data.
+/// To create an object which upwards its properties, we use `U`.
+/// To create an HTML element, we use `E`,
+/// and provide the tag, children, and event handlers.
+/// We then use the property as the label of a button.
+/// When the button is clicked, changing the label property automatically updates the button label.
+///
+/// Why is there no function to create a button directly,
+/// such as `UpButton(labelText, clickHandler)`?
+/// Upward maintains a compact surface area.
+/// If you want a convenience function, write one yourself.
+
 //===START
 config = U({ label: "Press me" });
 function change() { config.label = "Pressed!"; }
