+++ sam/Srt.js
@@ -1,5 +1,16 @@
+// sam/Srt.js
+// Sample for sorting.
+
 import {U, T, E} from '../src/Up';
+
 var dom, arr;
+
+
+/// ### Sorting
+///
+/// This example is nearly identical to the previous, just adding
+/// the `by` preposition to ensure the array is sorted.
+
 
 //===START
 arr = [3, 2, 1];
@@ -9,7 +20,7 @@ function add()    { arr . push(random()); }
 dom = E('div') . has( [
   E('div') . has(
     arr .
-      by(Number) . 
+      by(Number) .
       as(val => E('div') . has(T(val)))
   ),
   E('button') . has(T("Add")) . does({click: add})
