+++ tst/Ify.js
@@ -1,40 +1,39 @@
 // Unit tests for src/Ify.js.
 
 import {test, testGroup, skip} from '../src/Tst';
-import {assert} from '../src/Cha';
 import {noop, compose, swapify, argify, invertify, dropify} from '../src/Ify';
 
 export default testGroup(
   "module Ify (functional programming)",
   [
 
-    test("compose", () => {
+    test("compose", ({assert}) => {
       var addone = x => x + 1;
       var multiplytwo = x => x * 2;
       assert.equal(compose(addone, multiplytwo)(1), 3);
     }),
 
-    test("swapify", _ => {
+    test("swapify", ({assert}) => {
       var callback = sinon.spy();
       var proxy = swapify(callback);
       proxy(1, 2);
       assert(callback.calledWith(2, 1));
     }),
 
-    test("argify", _ => {
+    test("argify", ({assert}) => {
       var callback = sinon.spy();
       var proxy = argify(callback, 1);
       proxy(2);
       assert(callback.calledWith(1, 2));
     }),
 
-    test("invertify", _ => {
+    test("invertify", ({assert}) => {
       var callback = sinon.stub().returns(true);
       var proxy = invertify(callback);
       assert(!proxy());
     }),
 
-    test("dropify", _ => {
+    test("dropify", ({assert}) => {
       var callback = sinon.spy();
       var proxy = dropify(callback);
       proxy(1, 2);
